resume_agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.4
python-multipart>=0.0.12
python-dotenv>=1.0.1
pydantic[email]>=2.9.2
crewai[litellm]>=0.175.0
python-docx>=1.1.2
reportlab>=4.2.5
